+++ test/integration/ui/isZooming.test.ts
@@ -1,8 +1,8 @@
-import {browser} from '../../util/browser';
-import {Map} from '../../ui/map';
-import {DOM} from '../../util/dom';
-import simulate from '../../../test/unit/lib/simulate_interaction';
-import {beforeMapTest} from '../../util/test/util';
+import {browser} from '../../../src/util/browser';
+import {Map} from '../../../src/ui/map';
+import {DOM} from '../../../src/util/dom';
+import simulate from '../../unit/lib/simulate_interaction';
+import {beforeMapTest} from '../../../src/util/test/util';
 
 function createMap() {
     return new Map({style: '', container: DOM.create('div', '', window.document.body)});
